Add an HTTP server running a web UI with a button for each command
Now the binary has two HTTP servers: one for the API, which requires
token validation, and one for the web UI, which does not. They run on
different ports and only the API server's port should be exposed to
the outside world.

This also changes the default port for the API server from 8080 to 9090.
The default port for the web UI is 8080.